requirements.txt
fastapi
uvicorn
pillow
python-dotenv
google-generativeai
python-multipart
streamlit
openai
